Use struct with state not package
